feat: update duration format in MovieDetailsState

+++ presentation/src/main/java/com/moscow/cineverse/screen/myratings/MyRatingsContent.kt
@@ -21,6 +21,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.paging.LoadState
 import androidx.paging.compose.LazyPagingItems
+import com.moscow.cineverse.common_ui_state.MediaItemUiState
 import com.moscow.cineverse.component.EmptyState
 import com.moscow.cineverse.component.MediaPosterCard
 import com.moscow.cineverse.component.NoInternetScreen
@@ -29,7 +30,7 @@ import com.moscow.cineverse.designSystem.component.card.InfoCard
 import com.moscow.cineverse.designSystem.component.indicator.MovieCircularProgressBar
 import com.moscow.cineverse.designSystem.component.wrapper.MovieScaffold
 import com.moscow.cineverse.designSystem.theme.Theme
-import com.moscow.cineverse.details.common.MovieRatingBar
+import com.moscow.cineverse.screen.details.common.MovieRatingBar
 import com.moscow.cineverse.screen.explore.component.ExploreTabsSection
 import com.moscow.cineverse.utlis.ViewMode
 import com.moscow.cinverse.presentation.R
@@ -143,7 +144,7 @@ private fun MyRatingsList(
                             color = Theme.colors.shade.primary
                         )
                         MovieRatingBar(
-                            mediaItem.rating,
+                            mediaItem.rating.toFloat(),
                             onRatingChanged = {},
                         )
                     }
